Make sure we have a record before trying to delete

+++ BrianJacobs/Scheduler/Site/Controllers/AppointmentController.php
@@ -176,7 +176,10 @@ class AppointmentController extends BaseController {
 				})->first();
 
 				// Remove the user appointment
-				$userAppt->forceDelete();
+				if ($userAppt)
+				{
+					$userAppt->forceDelete();
+				}
 			}
 		}
 	}
